add 项目价格编辑接口

+++ common/src/main/java/com/ligg/common/entity/ProjectEntity.java
@@ -19,6 +19,6 @@ public class ProjectEntity {
 //    @JsonFormat(shape = JsonFormat.Shape.STRING)
     private Integer projectId;
     private String projectName;
-    private Float projectPrice;// 项目价格
+    private Double projectPrice;// 项目价格
     private LocalDateTime projectCreatedAt;// 创建时间
 }
